Add a utility function module
